Cleans up code and adds uname spoofing
Addresses minor code style issues, such as whitespace inconsistencies and parenthesis usage, for improved readability.

Introduces uname spoofing for the "netbpfload" process, setting the release version to "6.6.40". This allows specific processes to report a different kernel version than the actual one. This change is gated by a PID check to ensure it only applies to the intended process instance.

Signed-off-by: Zears14 <79321017+Zears14@users.noreply.github.com>